casper_name_contracts::contracts
Module controller

Structs

BaseController

Base for all controllers. It handles access control, treasury and signer public key.

BaseControllerContractRef

BaseController Contract Ref.

BaseControllerHostRef

BaseController Host Ref.

Controller

Controller smart contract. It handles payments and talks to the Registrar Contract.

ControllerContractRef

Controller Contract Ref.

ControllerHostRef

Controller Host Ref.

ControllerInitArgs

Controller contract constructor arguments.

PaymentFulfilled

Event with the payment information.

SignerPublicKeyChanged

Event emitted when the signer public key is changed.

TreasuryAddressChanged

Event emitted when the treasury address is changed.

Enums

ControllerError

Controller errors.