casper_name_contracts::contracts::controller

Struct BaseControllerHostRef

pub struct BaseControllerHostRef { /* private fields */ }

BaseController Host Ref.

Implementations

impl BaseControllerHostRef

pub fn has_role(&self, role: &Role, address: &Address) -> bool

Delegated. See self.access_control.has_role() for details.

pub fn grant_role(&mut self, role: &Role, address: &Address)

Delegated. See self.access_control.grant_role() for details.

pub fn revoke_role(&mut self, role: &Role, address: &Address)

Delegated. See self.access_control.revoke_role() for details.

pub fn is_paused(&self) -> bool

Delegated. See self.pausable.is_paused() for details.

pub fn require_not_paused(&self)

Delegated. See self.pausable.require_not_paused() for details.

impl BaseControllerHostRef

pub fn try_has_role(&self, role: &Role, address: &Address) -> OdraResult<bool>

Delegated. See self.access_control.has_role() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_grant_role( &mut self, role: &Role, address: &Address, ) -> OdraResult<()>

Delegated. See self.access_control.grant_role() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_revoke_role( &mut self, role: &Role, address: &Address, ) -> OdraResult<()>

Delegated. See self.access_control.revoke_role() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_is_paused(&self) -> OdraResult<bool>

Delegated. See self.pausable.is_paused() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_require_not_paused(&self) -> OdraResult<()>

Delegated. See self.pausable.require_not_paused() for details. Does not fail in case of error, returns odra::OdraResult instead.