Struct casper_name_contracts::contracts::controller::ControllerContractRef

pub struct ControllerContractRef { /* private fields */ }

Controller Contract Ref.

Implementations

impl ControllerContractRef

pub fn init(&mut self, registrar: Address, signer: PublicKey, treasury: Address)

Initializes the controller with the registrar contract address, the signer public key and the treasury address.

pub fn buy(&mut self, voucher: PaymentVoucher, signature: Bytes)

Payable. Buys new name tokens.

pub fn renew(&mut self, voucher: RenewalPaymentVoucher, signature: Bytes)

Payable. Renews name tokens.

pub fn buy_and_renew( &mut self, payment_voucher: PaymentVoucher, payment_signature: Bytes, renewal_voucher: RenewalPaymentVoucher, renewal_signature: Bytes )

Payable. Buys new name tokens and renews existing ones.

pub fn resolve(&self, full_domain: String) -> Option<Address>

Try to resolve a full domain name to an address.

pub fn has_role(&self, role: &Role, address: &Address) -> bool

Delegated. See self.controller.has_role() for details.

pub fn grant_role(&mut self, role: &Role, address: &Address)

Delegated. See self.controller.grant_role() for details.

pub fn revoke_role(&mut self, role: &Role, address: &Address)

Delegated. See self.controller.revoke_role() for details.

pub fn set_signer_public_key(&mut self, signer: PublicKey)

Delegated. See self.controller.set_signer_public_key() for details.

pub fn set_treasury(&mut self, treasury: Address)

Delegated. See self.controller.set_treasury() for details.

pub fn signer_public_key(&self) -> PublicKey

Delegated. See self.controller.signer_public_key() for details.