Struct casper_name_contracts::contracts::marketplace::SecondaryMarketHostRef

pub struct SecondaryMarketHostRef { /* private fields */ }

SecondaryMarket Host Ref.

Implementations

impl SecondaryMarketHostRef

pub fn init( &mut self, signer: PublicKey, treasury: Address, name_token: Address )

Initializes the secondary market with the signer public key, the treasury address and the name token contract address.

pub fn buy(&mut self, voucher: SecondarySaleVoucher, signature: Bytes)

Payable. Buys name tokens from the secondary market.

pub fn has_role(&self, role: &Role, address: &Address) -> bool

Delegated. See self.controller.has_role() for details.

pub fn grant_role(&mut self, role: &Role, address: &Address)

Delegated. See self.controller.grant_role() for details.

pub fn revoke_role(&mut self, role: &Role, address: &Address)

Delegated. See self.controller.revoke_role() for details.

pub fn set_signer_public_key(&mut self, signer: PublicKey)

Delegated. See self.controller.set_signer_public_key() for details.

pub fn set_treasury(&mut self, treasury: Address)

Delegated. See self.controller.set_treasury() for details.

pub fn signer_public_key(&self) -> PublicKey

Delegated. See self.controller.signer_public_key() for details.

impl SecondaryMarketHostRef

pub fn try_init( &mut self, signer: PublicKey, treasury: Address, name_token: Address ) -> OdraResult<()>

Initializes the secondary market with the signer public key, the treasury address and the name token contract address. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_buy( &mut self, voucher: SecondarySaleVoucher, signature: Bytes ) -> OdraResult<()>

Payable. Buys name tokens from the secondary market. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_has_role(&self, role: &Role, address: &Address) -> OdraResult<bool>

Delegated. See self.controller.has_role() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_grant_role( &mut self, role: &Role, address: &Address ) -> OdraResult<()>

Delegated. See self.controller.grant_role() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_revoke_role( &mut self, role: &Role, address: &Address ) -> OdraResult<()>

Delegated. See self.controller.revoke_role() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_set_signer_public_key(&mut self, signer: PublicKey) -> OdraResult<()>

Delegated. See self.controller.set_signer_public_key() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_set_treasury(&mut self, treasury: Address) -> OdraResult<()>

Delegated. See self.controller.set_treasury() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_signer_public_key(&self) -> OdraResult<PublicKey>

Delegated. See self.controller.signer_public_key() for details. Does not fail in case of error, returns odra::OdraResult instead.