Struct casper_name_contracts::contracts::name_token::NameToken

pub struct NameToken { /* private fields */ }

NameToken contract. It is a CEP78 token with additional functionalities.

Implementations

impl NameToken

pub fn init(&mut self, name: String, symbol: String)

Initializes CEP78 with the given name and symbol.

pub fn token_exists(&self, token_hash: &String) -> bool

Checks if a token with the given hash exists.

pub fn burn(&mut self, token_id: Maybe<u64>, token_hash: Maybe<String>)

Only admin. Burns the token with the given hash.

pub fn admin_transfer(&mut self, recipient: Address, token_hashes: Vec<String>)

Only admin. Transfer tokens to the given recipient.

pub fn transfer( &mut self, token_id: Maybe<u64>, token_hash: Maybe<String>, source_key: Address, target_key: Address )

Transfer token.

pub fn set_token_metadata( &mut self, token_id: Maybe<u64>, token_hash: Maybe<String>, token_meta_data: String )

Only admin. Set token’s metadata.

pub fn metadata_by_hash(&self, token_hash: String) -> String

Return the metadata of the token with the given hash.

pub fn resolver(&self, token_hash: String) -> Option<Address>

Return the resolver of the token with the given hash.

pub fn set_resolver(&mut self, token_hash: String, resolver: Address)

Token owner only. Set the resolver of the token with the given hash.

pub fn assert_is_owner(&self, token_hash: &String, address: Address)

Check if the address is the owner of the token with the given hash.

pub fn is_token_valid(&self, token_hash: &String) -> bool

Check if the token with the given hash is expired.

pub fn set_default_resolver(&mut self, resolver: Address)

Only admin. Set the default resolver.

pub fn get_default_resolver(&self) -> Address

Get the default resolver.

pub fn get_collection_name(&self) -> String

Delegated. See self.token.get_collection_name() for details.

pub fn get_collection_symbol(&self) -> String

Delegated. See self.token.get_collection_symbol() for details.

pub fn set_variables( &mut self, allow_minting: Maybe<bool>, acl_whitelist: Maybe<Vec<Address>>, operator_burn_mode: Maybe<bool> )

Delegated. See self.token.set_variables() for details.

pub fn mint( &mut self, token_owner: Address, token_meta_data: String, token_hash: Maybe<String> )

Delegated. See self.token.mint() for details.

pub fn approve( &mut self, spender: Address, token_id: Maybe<u64>, token_hash: Maybe<String> )

Delegated. See self.token.approve() for details.

pub fn revoke(&mut self, token_id: Maybe<u64>, token_hash: Maybe<String>)

Delegated. See self.token.revoke() for details.

pub fn set_approval_for_all(&mut self, approve_all: bool, operator: Address)

Delegated. See self.token.set_approval_for_all() for details.

pub fn is_approved_for_all( &mut self, token_owner: Address, operator: Address ) -> bool

Delegated. See self.token.is_approved_for_all() for details.

pub fn owner_of( &self, token_id: Maybe<u64>, token_hash: Maybe<String> ) -> Address

Delegated. See self.token.owner_of() for details.

pub fn get_approved( &mut self, token_id: Maybe<u64>, token_hash: Maybe<String> ) -> Option<Address>

Delegated. See self.token.get_approved() for details.

pub fn metadata( &self, token_id: Maybe<u64>, token_hash: Maybe<String> ) -> String

Delegated. See self.token.metadata() for details.

pub fn balance_of(&mut self, token_owner: Address) -> u64

Delegated. See self.token.balance_of() for details.

pub fn register_owner(&mut self, token_owner: Maybe<Address>) -> String

Delegated. See self.token.register_owner() for details.

impl NameToken

pub fn wrapped_metadata(&self, token_hash: &str) -> NameTokenMetadata