Struct casper_name_contracts::contracts::name_token::NameTokenHostRef

pub struct NameTokenHostRef { /* private fields */ }

NameToken Host Ref.

Implementations

impl NameTokenHostRef

pub fn init(&mut self, name: String, symbol: String)

Initializes CEP78 with the given name and symbol.

pub fn token_exists(&self, token_hash: &String) -> bool

Checks if a token with the given hash exists.

pub fn burn(&mut self, token_id: Maybe<u64>, token_hash: Maybe<String>)

Only admin. Burns the token with the given hash.

pub fn admin_transfer(&mut self, recipient: Address, token_hashes: Vec<String>)

Only admin. Transfer tokens to the given recipient.

pub fn transfer( &mut self, token_id: Maybe<u64>, token_hash: Maybe<String>, source_key: Address, target_key: Address )

Transfer token.

pub fn set_token_metadata( &mut self, token_id: Maybe<u64>, token_hash: Maybe<String>, token_meta_data: String )

Only admin. Set token’s metadata.

pub fn metadata_by_hash(&self, token_hash: String) -> String

Return the metadata of the token with the given hash.

pub fn resolver(&self, token_hash: String) -> Option<Address>

Return the resolver of the token with the given hash.

pub fn set_resolver(&mut self, token_hash: String, resolver: Address)

Token owner only. Set the resolver of the token with the given hash.

pub fn assert_is_owner(&self, token_hash: &String, address: Address)

Check if the address is the owner of the token with the given hash.

pub fn is_token_valid(&self, token_hash: &String) -> bool

Check if the token with the given hash is expired.

pub fn set_default_resolver(&mut self, resolver: Address)

Only admin. Set the default resolver.

pub fn get_default_resolver(&self) -> Address

Get the default resolver.

pub fn get_collection_name(&self) -> String

Delegated. See self.token.get_collection_name() for details.

pub fn get_collection_symbol(&self) -> String

Delegated. See self.token.get_collection_symbol() for details.

pub fn set_variables( &mut self, allow_minting: Maybe<bool>, acl_whitelist: Maybe<Vec<Address>>, operator_burn_mode: Maybe<bool> )

Delegated. See self.token.set_variables() for details.

pub fn mint( &mut self, token_owner: Address, token_meta_data: String, token_hash: Maybe<String> )

Delegated. See self.token.mint() for details.

pub fn approve( &mut self, spender: Address, token_id: Maybe<u64>, token_hash: Maybe<String> )

Delegated. See self.token.approve() for details.

pub fn revoke(&mut self, token_id: Maybe<u64>, token_hash: Maybe<String>)

Delegated. See self.token.revoke() for details.

pub fn set_approval_for_all(&mut self, approve_all: bool, operator: Address)

Delegated. See self.token.set_approval_for_all() for details.

pub fn is_approved_for_all( &mut self, token_owner: Address, operator: Address ) -> bool

Delegated. See self.token.is_approved_for_all() for details.

pub fn owner_of( &self, token_id: Maybe<u64>, token_hash: Maybe<String> ) -> Address

Delegated. See self.token.owner_of() for details.

pub fn get_approved( &mut self, token_id: Maybe<u64>, token_hash: Maybe<String> ) -> Option<Address>

Delegated. See self.token.get_approved() for details.

pub fn metadata( &self, token_id: Maybe<u64>, token_hash: Maybe<String> ) -> String

Delegated. See self.token.metadata() for details.

pub fn balance_of(&mut self, token_owner: Address) -> u64

Delegated. See self.token.balance_of() for details.

pub fn register_owner(&mut self, token_owner: Maybe<Address>) -> String

Delegated. See self.token.register_owner() for details.

impl NameTokenHostRef

pub fn try_init(&mut self, name: String, symbol: String) -> OdraResult<()>

Initializes CEP78 with the given name and symbol. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_token_exists(&self, token_hash: &String) -> OdraResult<bool>

Checks if a token with the given hash exists. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_burn( &mut self, token_id: Maybe<u64>, token_hash: Maybe<String> ) -> OdraResult<()>

Only admin. Burns the token with the given hash. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_admin_transfer( &mut self, recipient: Address, token_hashes: Vec<String> ) -> OdraResult<()>

Only admin. Transfer tokens to the given recipient. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_transfer( &mut self, token_id: Maybe<u64>, token_hash: Maybe<String>, source_key: Address, target_key: Address ) -> OdraResult<()>

Transfer token. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_set_token_metadata( &mut self, token_id: Maybe<u64>, token_hash: Maybe<String>, token_meta_data: String ) -> OdraResult<()>

Only admin. Set token’s metadata. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_metadata_by_hash(&self, token_hash: String) -> OdraResult<String>

Return the metadata of the token with the given hash. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_resolver(&self, token_hash: String) -> OdraResult<Option<Address>>

Return the resolver of the token with the given hash. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_set_resolver( &mut self, token_hash: String, resolver: Address ) -> OdraResult<()>

Token owner only. Set the resolver of the token with the given hash. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_assert_is_owner( &self, token_hash: &String, address: Address ) -> OdraResult<()>

Check if the address is the owner of the token with the given hash. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_is_token_valid(&self, token_hash: &String) -> OdraResult<bool>

Check if the token with the given hash is expired. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_set_default_resolver(&mut self, resolver: Address) -> OdraResult<()>

Only admin. Set the default resolver. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_get_default_resolver(&self) -> OdraResult<Address>

Get the default resolver. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_get_collection_name(&self) -> OdraResult<String>

Delegated. See self.token.get_collection_name() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_get_collection_symbol(&self) -> OdraResult<String>

Delegated. See self.token.get_collection_symbol() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_set_variables( &mut self, allow_minting: Maybe<bool>, acl_whitelist: Maybe<Vec<Address>>, operator_burn_mode: Maybe<bool> ) -> OdraResult<()>

Delegated. See self.token.set_variables() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_mint( &mut self, token_owner: Address, token_meta_data: String, token_hash: Maybe<String> ) -> OdraResult<()>

Delegated. See self.token.mint() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_approve( &mut self, spender: Address, token_id: Maybe<u64>, token_hash: Maybe<String> ) -> OdraResult<()>

Delegated. See self.token.approve() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_revoke( &mut self, token_id: Maybe<u64>, token_hash: Maybe<String> ) -> OdraResult<()>

Delegated. See self.token.revoke() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_set_approval_for_all( &mut self, approve_all: bool, operator: Address ) -> OdraResult<()>

Delegated. See self.token.set_approval_for_all() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_is_approved_for_all( &mut self, token_owner: Address, operator: Address ) -> OdraResult<bool>

Delegated. See self.token.is_approved_for_all() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_owner_of( &self, token_id: Maybe<u64>, token_hash: Maybe<String> ) -> OdraResult<Address>

Delegated. See self.token.owner_of() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_get_approved( &mut self, token_id: Maybe<u64>, token_hash: Maybe<String> ) -> OdraResult<Option<Address>>

Delegated. See self.token.get_approved() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_metadata( &self, token_id: Maybe<u64>, token_hash: Maybe<String> ) -> OdraResult<String>

Delegated. See self.token.metadata() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_balance_of(&mut self, token_owner: Address) -> OdraResult<u64>

Delegated. See self.token.balance_of() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_register_owner( &mut self, token_owner: Maybe<Address> ) -> OdraResult<String>

Delegated. See self.token.register_owner() for details. Does not fail in case of error, returns odra::OdraResult instead.