casper_name_contracts::contracts::registrar
Struct RegistrarContractRef

pub struct RegistrarContractRef { /* private fields */ }

Registrar Contract Ref.

Implementations

impl RegistrarContractRef

pub fn init(&mut self, name_token: Address)

Initializes the registrar with the name token contract address.

pub fn pause(&mut self)

Temporarily stops the contract.

pub fn unpause(&mut self)

Returns to normal operation.

pub fn grace_period(&self) -> u64

Returns the grace period.

pub fn resolve(&self, full_domain: String) -> Option<Address>

Try to resolve a full domain name to an address.

pub fn expire(&mut self, token_ids: Vec<U256>)

Expire a list of tokens if they are expired.

pub fn set_grace_period(&mut self, period: u64)

Admin only. Sets the grace period.

pub fn admin_transfer(&mut self, new_owner: Address, token_ids: Vec<U256>)

Admin only. Transfer ownership of a list of tokens.

pub fn admin_burn(&mut self, token_ids: Vec<U256>)

Admin only. Burn a list of tokens.

pub fn admin_prolong(&mut self, tokens: Vec<TokenRenewalInfo>)

Admin only. Prolong the expiration date of a list of tokens.

pub fn admin_register(&mut self, names: Vec<NameMintInfo>)

Admin only. Register a list of tokens.

pub fn admin_prolong_and_register( &mut self, renewal_tokens: Vec<TokenRenewalInfo>, new_tokens: Vec<NameMintInfo>, )

Admin only. Prolong the expiration date of a list of tokens and register a list of tokens.

pub fn controller_prolong(&mut self, voucher: RenewalVoucher)

Controller only. Prolong the expiration date of a list of tokens.

pub fn controller_register(&mut self, voucher: TokenizationVoucher)

Controller only. Register a list of tokens.

pub fn controller_prolong_and_register( &mut self, renewal_voucher: RenewalVoucher, tokenization_voucher: TokenizationVoucher, )

Controller only. Prolong the expiration date of a list of tokens and register a list of tokens.

pub fn has_role(&self, role: &Role, address: &Address) -> bool

Delegated. See self.access_control.has_role() for details.

pub fn grant_role(&mut self, role: &Role, address: &Address)

Delegated. See self.access_control.grant_role() for details.

pub fn revoke_role(&mut self, role: &Role, address: &Address)

Delegated. See self.access_control.revoke_role() for details.

pub fn is_paused(&self) -> bool

Delegated. See self.pauseable.is_paused() for details.