casper_name_contracts::contracts::resolver

Struct ResolverHostRef

pub struct ResolverHostRef { /* private fields */ }

[Resolver] Host Ref.

Implementations

impl ResolverHostRef

pub fn init(&mut self, name_token: Address)

pub fn set_name_token(&mut self, name_token: Address)

pub fn set_resolution(&mut self, full_domain: String, address: Option<Address>)

pub fn resolve(&self, full_domain: String) -> Option<Address>

pub fn invalidate_resolutions(&mut self, token_id: U256)

impl ResolverHostRef

pub fn try_init(&mut self, name_token: Address) -> OdraResult<()>

Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_set_name_token(&mut self, name_token: Address) -> OdraResult<()>

Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_set_resolution( &mut self, full_domain: String, address: Option<Address>, ) -> OdraResult<()>

Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_resolve(&self, full_domain: String) -> OdraResult<Option<Address>>

Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_invalidate_resolutions(&mut self, token_id: U256) -> OdraResult<()>

Does not fail in case of error, returns odra::OdraResult instead.