casper_name_contracts

Module data_structures

Structs

NameMintInfo

Basic minting information for a name token.

NameTokenMetadata

Metadata associated with a name token.

NameTransferInfo

Pair of a label and owner address.

PaymentInfo

Information about a payment.

PaymentVoucher

Information about a payment and a list of NameMintInfo structs.

RenewalPaymentVoucher

Voucher for renewing multiple name tokens, plus payment information.

RenewalVoucher

Voucher for renewing multiple name tokens.

SecondarySaleVoucher

Information about a payment and a list of NameTransferInfo structs.

TokenRenewalInfo

Renewal information with new expiration time.

TokenizationVoucher

List of NameMintInfo structs and the expiration time of the voucher.

Enums

NameTokenError

Errors that can occur while working with name tokens.

Traits

ExpirableVoucher

Payment