casper_name_contracts::contracts::name_token
Struct NameTokenHostRef

pub struct NameTokenHostRef { /* private fields */ }

NameToken Host Ref.

Implementations

impl NameTokenHostRef

pub fn init(&mut self, name: String, symbol: String, max_supply: u64)

Initializes CEP95 with the given name and symbol.

pub fn token_exists(&self, token_id: U256) -> bool

pub fn mint( &mut self, recipient: Address, token_id: U256, token_metadata: Vec<(String, String)>, )

pub fn burn(&mut self, token_id: U256)

pub fn admin_transfer(&mut self, recipient: Address, token_ids: Vec<U256>)

pub fn transfer_from(&mut self, from: Address, to: Address, token_id: U256)

pub fn set_token_metadata( &mut self, token_id: U256, token_metadata: Vec<(String, String)>, )

pub fn resolver(&self, token_id: U256) -> Option<Address>

pub fn set_resolver(&mut self, token_id: U256, resolver: Address)

pub fn assert_is_owner(&self, token_id: U256, address: Address)

pub fn is_token_valid(&self, token_id: U256) -> bool

pub fn set_default_resolver(&mut self, resolver: Address)

Only admin. Set the default resolver.

pub fn get_default_resolver(&self) -> Address

Get the default resolver.

pub fn whitelist(&mut self, address: Address)

pub fn revoke_whitelist(&mut self, address: Address)

pub fn name(&self) -> String

Delegated. See self.token.name() for details.

pub fn symbol(&self) -> String

Delegated. See self.token.symbol() for details.

pub fn balance_of(&self, owner: Address) -> U256

Delegated. See self.token.balance_of() for details.

pub fn owner_of(&self, token_id: U256) -> Option<Address>

Delegated. See self.token.owner_of() for details.

pub fn safe_transfer_from( &mut self, from: Address, to: Address, token_id: U256, data: Option<Bytes>, )

Delegated. See self.token.safe_transfer_from() for details.

pub fn approve(&mut self, spender: Address, token_id: U256)

Delegated. See self.token.approve() for details.

pub fn revoke_approval(&mut self, token_id: U256)

Delegated. See self.token.revoke_approval() for details.

pub fn approved_for(&self, token_id: U256) -> Option<Address>

Delegated. See self.token.approved_for() for details.

pub fn approve_for_all(&mut self, operator: Address)

Delegated. See self.token.approve_for_all() for details.

pub fn revoke_approval_for_all(&mut self, operator: Address)

Delegated. See self.token.revoke_approval_for_all() for details.

pub fn is_approved_for_all(&self, owner: Address, operator: Address) -> bool

Delegated. See self.token.is_approved_for_all() for details.

pub fn token_metadata(&self, token_id: U256) -> Vec<(String, String)>

Delegated. See self.token.token_metadata() for details.

pub fn get_owner(&self) -> Address

Delegated. See self.ownable.get_owner() for details.

pub fn get_pending_owner(&self) -> Option<Address>

Delegated. See self.ownable.get_pending_owner() for details.

pub fn transfer_ownership(&mut self, new_owner: &Address)

Delegated. See self.ownable.transfer_ownership() for details.

pub fn accept_ownership(&mut self)

Delegated. See self.ownable.accept_ownership() for details.

pub fn renounce_ownership(&mut self)

Delegated. See self.ownable.renounce_ownership() for details.

impl NameTokenHostRef

pub fn try_init( &mut self, name: String, symbol: String, max_supply: u64, ) -> OdraResult<()>

Initializes CEP95 with the given name and symbol. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_token_exists(&self, token_id: U256) -> OdraResult<bool>

Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_mint( &mut self, recipient: Address, token_id: U256, token_metadata: Vec<(String, String)>, ) -> OdraResult<()>

Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_burn(&mut self, token_id: U256) -> OdraResult<()>

Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_admin_transfer( &mut self, recipient: Address, token_ids: Vec<U256>, ) -> OdraResult<()>

Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_transfer_from( &mut self, from: Address, to: Address, token_id: U256, ) -> OdraResult<()>

Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_set_token_metadata( &mut self, token_id: U256, token_metadata: Vec<(String, String)>, ) -> OdraResult<()>

Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_resolver(&self, token_id: U256) -> OdraResult<Option<Address>>

Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_set_resolver( &mut self, token_id: U256, resolver: Address, ) -> OdraResult<()>

Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_assert_is_owner( &self, token_id: U256, address: Address, ) -> OdraResult<()>

Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_is_token_valid(&self, token_id: U256) -> OdraResult<bool>

Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_set_default_resolver(&mut self, resolver: Address) -> OdraResult<()>

Only admin. Set the default resolver. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_get_default_resolver(&self) -> OdraResult<Address>

Get the default resolver. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_whitelist(&mut self, address: Address) -> OdraResult<()>

Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_revoke_whitelist(&mut self, address: Address) -> OdraResult<()>

Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_name(&self) -> OdraResult<String>

Delegated. See self.token.name() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_symbol(&self) -> OdraResult<String>

Delegated. See self.token.symbol() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_balance_of(&self, owner: Address) -> OdraResult<U256>

Delegated. See self.token.balance_of() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_owner_of(&self, token_id: U256) -> OdraResult<Option<Address>>

Delegated. See self.token.owner_of() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_safe_transfer_from( &mut self, from: Address, to: Address, token_id: U256, data: Option<Bytes>, ) -> OdraResult<()>

Delegated. See self.token.safe_transfer_from() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_approve( &mut self, spender: Address, token_id: U256, ) -> OdraResult<()>

Delegated. See self.token.approve() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_revoke_approval(&mut self, token_id: U256) -> OdraResult<()>

Delegated. See self.token.revoke_approval() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_approved_for(&self, token_id: U256) -> OdraResult<Option<Address>>

Delegated. See self.token.approved_for() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_approve_for_all(&mut self, operator: Address) -> OdraResult<()>

Delegated. See self.token.approve_for_all() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_revoke_approval_for_all( &mut self, operator: Address, ) -> OdraResult<()>

Delegated. See self.token.revoke_approval_for_all() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_is_approved_for_all( &self, owner: Address, operator: Address, ) -> OdraResult<bool>

Delegated. See self.token.is_approved_for_all() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_token_metadata( &self, token_id: U256, ) -> OdraResult<Vec<(String, String)>>

Delegated. See self.token.token_metadata() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_get_owner(&self) -> OdraResult<Address>

Delegated. See self.ownable.get_owner() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_get_pending_owner(&self) -> OdraResult<Option<Address>>

Delegated. See self.ownable.get_pending_owner() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_transfer_ownership(&mut self, new_owner: &Address) -> OdraResult<()>

Delegated. See self.ownable.transfer_ownership() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_accept_ownership(&mut self) -> OdraResult<()>

Delegated. See self.ownable.accept_ownership() for details. Does not fail in case of error, returns odra::OdraResult instead.

pub fn try_renounce_ownership(&mut self) -> OdraResult<()>

Delegated. See self.ownable.renounce_ownership() for details. Does not fail in case of error, returns odra::OdraResult instead.