casper_name_contracts::contracts

Module resolver

Structs

DefaultResolver

Default Resolver smart contract. It handles the resolution of domain names to addresses.

DefaultResolverContractRef

DefaultResolver Contract Ref.

DefaultResolverHostRef

DefaultResolver Host Ref.

DefaultResolverInitArgs

DefaultResolver contract constructor arguments.

NameTokenAddressChanged

Event emitted when the name token address is changed.

ResolutionChanged

Event emitted when a resolution is changed.

ResolutionCleared

Event emitted when a resolution is cleared.

ResolverContractRef

[Resolver] Contract Ref.

ResolverHostRef

[Resolver] Host Ref.

Enums

ResolverError